// packages/@keyspot/cli/src/logger.d.ts
export declare const log: {
    scanning: (msg: string) => void;
    detected: (msg: string) => void;
    vaulted: (msg: string) => void;
    clean: (msg: string) => void;
    info: (msg: string) => void;
    muted: (msg: string) => void;
    error: (msg: string) => void;
};
//# sourceMappingURL=logger.d.ts.map